pci-epf: allow cqid specifying for sq creation
The NVMe PCIe specification allows submission queues to share
completion queues, that is, they do not need to be mapped 1 to 1.

This patch allows us to specify completion queue ID from the host
command, such that it can be used to keep track of the CQID that the
SQID being created is mapped to.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>