nvmet: sq_create: allow passing a cq
When a submission queue is created, this allows us to pass in a
reference to the corresponding completion queue that the submission
queue is mapped to. This is particularly relevent for PCIe where
submission queues may share a single completion queue, instead of being
a one to one submission queue to completion queue (id based) mapping.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>