nvmet: allow completion queue sharing for PCI controllers
When a submission queue is created, this allows us to pass in a
reference to the corresponding completion queue that the submission
queue is mapped to. This is particularly relevant for PCI controllers
where many submission queues may share a single completion queue.

With submission queues now interanally referencing the respective
completion queue, reference counting is used to avoid situations where
an attempt to delete a completion queue that is actively mapped to
a submission queue.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>