nvmet: allow completion queue sharing for PCI controllers
When a submission queue is created, this patch allows passing a
reference to the corresponding completion queue that the submission
queue is to use. This is articularly relevant for PCI controllers
where many submission queues may share a single completion queue.

With submission queues now internally referencing the respective
completion queue, reference counting is used to avoid situations where
an attempt to delete a completion queue may occur, whilst it is actively
used by a submission queue.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>