nvmet: cq: prepare for completion queue sharing
This is a preparatory patch that sets up allowing for completion queue
sharing by submission queues.

For PCI, the NVMe specification allows submission queues to share
completion queues, however, this is not supported in the current
implementation. To support queue sharing, reference counting
completion queues is required, this patch adds a `refcount_t` to the
`nvmet_cq` struct coupled with respective `nvmet_cq_init()` and
`nvmet_cq_destroy()` functions.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>