nvmet: support completion queue sharing
The NVMe specification for PCI allows for completion queues to be shared
by submission queues.

This patch allows a submission queue to keep track of the completion queue
it's using with reference counting. As such, it can be ensured that a completion
queues is not deleted while a submission queue is actively using it.

Completion queue sharing is now enabled for the `nvmet pci-epf` driver.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>