nvmet: support completion queue sharing
The NVMe PCI transport specification allows for completion queues to be shared
by submission queues.

This patch allows a submission queue to keep track of the completion queue
it is using with reference counting. As such, it can be ensured that a
completion queue is not deleted while a submission queue is actively
using it.

Completion queue sharing is now enabled for the nvmet pci-epf driver.
For fabrics drivers, completion queue sharing is not enabled as it is
not possible as per the fabrics specification. However, this patch
ammends the fabrics drivers to corrently integrate the new API that supports
completion queue sharing.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>