pci-epf: nvme: clear internal state of CC enable.
When a controller is disabled, clear the Controller Configuration (CC)
enable (EN) bitfield. Such that when the host attempts to re-enable the
controller by writng `1` to the CC_EN field, our internal state can
correctly detect that this field transitioned from a `0` -> `1`.

In the currently implementation, this is missed and the controller will
not be re-enabled. Reproducible by host by issuing an `unbind`
followed by a `bind` operation (via sysfs). That is:

$echo "0000:01:00.0" | tee /sys/bus/pci/drivers/nvme/unbind
$echo "0000:01:00.0" | tee /sys/bus/pci/drivers/nvme/bind

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>