PCI: dwc: add a delay in reset slot handling
When testing a Rock5B RC with a ROCK5B EP implementing the NVMe PCI
EPF driver, after a sysfs initiated bus reset from the host, the
endpoint is not able to do outbound DMA to the host. The host reports
issues as below from the NVMe driver when binding.

```
[    1.541394] nvme nvme0: missing or invalid SUBNQN field.
```

Adding a small delay during reset handling allows for downstream
devices to be ready (for example in the case of a bus reset) before
link training is initiated. In the case of the Rock5b RC <-> Rock5b EP,
this fixes the bus reset related issue.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>