nvmet: add a helper function for cqid checking
This patch adds a new helper function nvmet_check_cmd_cqid. It is to be
used when parsing a host command to ensure that the specified completion
queue identifier (CQID) is not 0 (Admin queue ID). This is a check that
already occurs in the nvmet_execute_x functions prior to
nvmet_check_cqid.

With the addition of the helper function, CQID checks in the
nvmet_execute_x function can be removed, and instead simply just call
nvmet_check_cmd_cqid in place of nvmet_check_cqid.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>