PCI: dw-rockchip: Disable L1ss in EP mode
Disable L1 substates in EP mode.

I can run the pci_endpoint_test selftest successfully with ASPM enabled
with L0s and L1.

However, if we have the L1 substates (L1.1 and L1.2) enabled in the EP,
so ASPM can go to these states as well, all the tests in pci_endpoint_test
fail.

I have not tried to debug this any further.

Signed-off-by: Niklas Cassel <cassel@kernel.org>