net/handshake: Get negotiated tls record size limit
During a handshake, an endpoint may specify a maximum record size limit,
this limit is currently not visble to kernel. This patch adds support
for retrieving the record size limit. This is the first step in ensuring
that the kernel can respect the record size limit imposed by the
endpoint.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>